@@ -3,39 +3,40 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.32",
-    "@fortawesome/free-solid-svg-icons": "^5.15.1",
-    "@fortawesome/react-fontawesome": "^0.1.12",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.5.0",
-    "@testing-library/user-event": "^7.2.1",
-    "@types/is-hotkey": "^0.1.2",
-    "@types/jest": "^24.9.1",
-    "@types/lodash": "^4.14.162",
-    "@types/node": "^12.19.2",
-    "@types/react": "^16.9.54",
-    "@types/react-dom": "^16.9.9",
-    "@types/react-redux": "^7.1.9",
-    "@types/react-router-dom": "^5.1.6",
+    "@emotion/css": "^11.10.0",
+    "@fortawesome/fontawesome-svg-core": "^6.1.2",
+    "@fortawesome/free-solid-svg-icons": "^6.1.2",
+    "@fortawesome/react-fontawesome": "^0.2.0",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/user-event": "^13.5.0",
+    "@types/is-hotkey": "^0.1.7",
+    "@types/jest": "^27.5.2",
+    "@types/lodash": "^4.14.184",
+    "@types/node": "^16.11.47",
+    "@types/react": "^18.0.15",
+    "@types/react-dom": "^18.0.6",
+    "@types/react-redux": "^7.1.24",
+    "@types/react-router-dom": "^5.3.3",
     "@types/redux": "^3.6.0",
-    "@types/slate": "^0.47.7",
-    "@types/slate-react": "^0.22.9",
-    "date-fns": "^2.16.1",
-    "emotion": "^10.0.27",
-    "is-hotkey": "^0.1.6",
-    "lodash": "^4.17.20",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-dropdown": "^1.9.0",
-    "react-modal": "^3.11.2",
-    "react-redux": "^7.2.2",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.2",
-    "redux": "^4.0.5",
-    "slate": "^0.58.4",
-    "slate-history": "^0.58.4",
-    "slate-react": "^0.58.4",
-    "typescript": "^3.9.7"
+    "@types/slate": "^0.47.11",
+    "@types/slate-react": "^0.50.1",
+    "date-fns": "^2.29.2",
+    "is-hotkey": "^0.2.0",
+    "lodash": "^4.17.21",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-dropdown": "^1.10.0",
+    "react-modal": "^3.15.1",
+    "react-redux": "^8.0.2",
+    "react-router-dom": "^5.3.3",
+    "react-scripts": "5.0.1",
+    "redux": "^4.2.0",
+    "slate": "^0.82.0",
+    "slate-history": "^0.66.0",
+    "slate-react": "^0.82.0",
+    "typescript": "^4.7.4",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -44,7 +45,10 @@
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
-    "extends": "react-app"
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
   },
   "browserslist": {
     "production": [
@@ -59,7 +63,6 @@
     ]
   },
   "devDependencies": {
-    "@types/date-fns": "^2.6.0",
-    "@types/react-modal": "^3.10.6"
+    "@types/react-modal": "^3.13.1"
   }
 }
