@@ -1,26 +1,27 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import ReactDOM from "react-dom/client";
 import "./index.css";
 import App from "./App";
-import * as serviceWorker from "./serviceWorker";
 import { BrowserRouter } from "react-router-dom";
 import { Provider } from "react-redux";
 import configureStore from "./store/configureStore";
 import ErrorBoundary from "./components/ErrorBoundary";
-import ReactModal from "react-modal";
+import reportWebVitals from "./reportWebVitals";
 
-ReactDOM.render(
-  <Provider store={configureStore()}>
-    <BrowserRouter>
-      <ErrorBoundary>{[<App key="App" />]}</ErrorBoundary>
-    </BrowserRouter>
-  </Provider>,
-  document.getElementById("root")
+const root = ReactDOM.createRoot(
+  document.getElementById("root") as HTMLElement
+);
+root.render(
+  <React.StrictMode>
+    <Provider store={configureStore()}>
+      <BrowserRouter>
+        <ErrorBoundary>{[<App key="App" />]}</ErrorBoundary>
+      </BrowserRouter>
+    </Provider>
+  </React.StrictMode>
 );
 
-ReactModal.setAppElement("#root");
-
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: https://bit.ly/CRA-PWA
-serviceWorker.unregister();
+// If you want to start measuring performance in your app, pass a function
+// to log results (for example: reportWebVitals(console.log))
+// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
+reportWebVitals();
